requirements.txt
gunicorn
flask
mysql-connector-python
flask_mail
itsdangerous